Made a little test program for checking the find_index() function in timsort. It randomly generates an array of ints, sorts it using qsort, uses a linear search to find the correct insertion index, then calls find_index() from timsort. Then it compares those indices and also asserts the requirements hold.
This is not formal proof the function is correct, and not exhaustive, but it
has the advantage of generating failure cases that can be examined to figure
out where things went wrong.

This commit also includes a fix for the find_index() function.
The culprit was the calculation of the number of remaining elements.
Actually, the real cause was the helpful comment that explained clearly,
but incorrectly, how to calculate this. In short, we were usually 1 element
short because the next recursion would start at the next known candidate,
but includes that candidate.

Helpful diagram:

0 1 2 3 4 5 6 7 8 9 (10 elements)
* *   *       *

We're looking where to place 5 for example. The elements marked
with a * were examined. The next recursion should start at index 4,
and examine the range 4-5-6 (because we can omit 7-8-9)

The faulty reasoning however was:
if you have 10 elements, and the minimum insertion index so far is (4),
whe you recurse you have to examine "the rest", being 5-6-7-8-9.
And that then didn't include 4 hence the bug.