It turns out the same bug was not present in merge_hi(), but it was different nonetheless.
In my efforts to fix merging I assumed there was a problem with the bounding, while it was actually
the find_index() function. This is how this bug got introduced: it solved another bug somewhere else,
and I had not updated merge_hi() with the same bug, but instead fixed find_index(), revealing this one.

The original code was correct all along, and also much nicer.

Bounding the array by the insert index means just subtracting that from the count,
which reduces to 0 when the insert location is after the array, since then count and index
are equal.