In order to generate "Real World Data", or at least data with patterns that occur in practice I wanted a source of randomly generated integers that have a normal distribution.
This turns out to be non-trivial. Wikipedia led me to believe the happenin' thing
is to use the Ziggurat Algorithm. Instead of implementing this from the original
paper or Wikipedia I chose to use http://heliosphan.org/zigguratalgorithm/zigguratalgorithm.html by Colin Green
as a source and just port it to C.

This is just the init() function that calculates the various x/y coordinates.
Apart from implementing the rest, I'd also like to calculate/generate the constants somehow.