Added a function to calculate e^x with Taylor series.
So math.h already has exp() of course, but I wanted to see how a Taylor Series
would do in comparison. As it turns out, quite well and sometimes really poorly.

When the contribution to the result becomes < x I stop the expansion. The result
is usually very close, and sometimes the difference between exp() and taylor_exp() are
things like 32, 65536 and so on. Some investigation to be done.