taylor_exp3() where we keep a list of factors, sort it first (descendingly) and the perform the addition.
This has the advantage that we don't overflow that soon (since we don't keep total x^n and n! around).
But has the drawback of costing a lot of memory. The results are much better though, since we use our
floats better. Still getting magic numbers at some point though.