Added an exp3 that keeps separate track of the mantissa for improved accuracy, and Kahan algorithm.
The former is a bit more accurate, but still starts to be quite wrong around n=100. Mostly because the
larger terms dominate everything.

The latter implements the Kahan algorithm which does something very similar: keeping track of an error
(basically the lower order parts). It also diverges a lot around n=100.