Fixed the bug that would return the wrong buffer.
The actual sorting was ok, but since this uses an additional buffer
sometimes the results would be in the 'wrong' one. This didn't matter
as the wrapup merge phase would do the right thing.

The edge case was when the number of blocks was a power of 2.
Then every even power of 2 would have the result in 'base' and
every odd power would have it in 'buf'. Then no wrapup merge would
happen.

The solution was to set the output pointer to 'buf' (we already bail early if there was only 1 block)
and swap it around until be run out of bits in blocks_done. I find this very ugly but
a) it works
b) it's not really inefficient
c) my train stop is coming up in 30 seconds