Fixed the logic for Gallop Mode where finding the index of an element means to displace and not position.
So when we have array Left and Right and we find_index to determine 'where Left[0] belongs in Right'
and we get 4, that means that Right[4] is >= left and that Right[4] comes after Left[0] in the output.

It's sometimes confusing because something might come at Right[0] (meaning everythin in Right comes after)
or something comes at Right[ length(Right) ] meaning it comes after. One cannot really have an index
after the last_index.